postgres_to_es/etl/requirements.txt
elasticsearch==8.2.0
psycopg2-binary==2.9
pydantic==1.9.1